backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
PyMySQL==1.1.1
Pillow==12.1.1
python-dotenv==1.0.1
reportlab==4.2.2
djangorestframework-simplejwt==5.3.1
cryptography==44.0.2
